Restructure repo for multi-grammar support (postgres + future plpgsql)
Move grammar.js, src/, test/, bindings/, and build configs into postgres/
subdirectory following the tree-sitter-typescript multi-grammar pattern.
Shared generator scripts remain in script/ at the root.

Layout:
  postgres/         SQL grammar (grammar.js, src/, test/, bindings/)
  script/           shared generator (parse-gram-y.js, codegen.js, etc.)
  plpgsql/          (future) PL/pgSQL grammar

Updated: package.json paths, go.mod module path, CI workflow paths,
generate-grammar.js output path, harvest-conflicts.sh to run from
postgres/ directory.
