+++ package.json
@@ -2,8 +2,8 @@
   "name": "tree-sitter-postgres",
   "version": "0.1.0",
   "description": "PostgreSQL grammar for tree-sitter, generated from Postgres source",
-  "main": "bindings/node",
-  "types": "bindings/node",
+  "main": "postgres/bindings/node",
+  "types": "postgres/bindings/node",
   "keywords": [
     "parser",
     "tree-sitter",
@@ -12,12 +12,12 @@
     "sql"
   ],
   "files": [
-    "grammar.js",
-    "binding.gyp",
+    "postgres/grammar.js",
+    "postgres/binding.gyp",
+    "postgres/bindings/node/*",
+    "postgres/queries/*",
+    "postgres/src/**",
     "prebuilds/**",
-    "bindings/node/*",
-    "queries/*",
-    "src/**",
     "*.wasm"
   ],
   "license": "MIT",
@@ -39,19 +39,19 @@
   },
   "scripts": {
     "install": "node-gyp-build",
-    "prestart": "tree-sitter build --wasm",
-    "start": "tree-sitter playground",
-    "test": "node --test bindings/node/*_test.js"
+    "generate": "node script/generate-grammar.js",
+    "test": "cd postgres && tree-sitter test"
   },
   "tree-sitter": [
     {
       "scope": "source.sql",
+      "path": "postgres",
       "file-types": [
         "sql",
         "psql",
         "pgsql"
       ],
-      "highlights": "queries/highlights.scm",
+      "highlights": "postgres/queries/highlights.scm",
       "injection-regex": "^(sql|postgres|postgresql)$"
     }
   ]
