+++ postgres/bindings/node/binding.cc
