+++ postgres/go.mod
@@ -1,4 +1,4 @@
-module github.com/gmr/tree-sitter-postgres
+module github.com/gmr/tree-sitter-postgres/postgres
 
 go 1.23
