+++ postgres/test/corpus/ddl.txt
