+++ postgres/test/corpus/set_operations.txt
